+++ media/Changelog.md
@@ -5,8 +5,8 @@
 # 2.1-pre6
 <!-- @BuildNumber: 6 -->
 <!-- @Includes: [] -->
-<!-- @Date: "2025-02-08 13:49:00 KST" -->
-<!-- @Git: "7acb2a4" -->
+<!-- @Date: "2025-02-09 03:57:00 KST" -->
+<!-- @Git: "30ad4df" -->
 
 - Reduce reloading cost